musehu-file-decoder working

+++ saint/tempo-detection-eval/estimators/MiniBpmEstimator.cpp
@@ -1,7 +1,8 @@
 #include "MiniBpmEstimator.h"
 
 namespace saint {
-MiniBpmEstimator::MiniBpmEstimator(int sampleRate) : _miniBpm(sampleRate) {}
+MiniBpmEstimator::MiniBpmEstimator(int sampleRate)
+    : _miniBpm(static_cast<float>(sampleRate)) {}
 
 void MiniBpmEstimator::process(float *audio, size_t len) {
   _miniBpm.process(audio, static_cast<int>(len));
